+ automatic triggering refactored + thresholds from sim parameters removed

+++ source/EngineGpuKernels/Object.cuh
@@ -88,7 +88,7 @@ struct TransmitterType
 struct ConstructorType
 {
     // Properties
-    uint32_t autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
+    uint8_t autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
     uint32_t constructionActivationTime;
 
     // Genome data
@@ -113,7 +113,7 @@ struct ConstructorType
 
 struct SensorType
 {
-    int autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
+    uint8_t autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
     float minDensity;
     int8_t minRange;          //< 0 = no restriction
     int8_t maxRange;          //< 0 = no restriction
